money type

+++ Homework1_SQL/1.sql
@@ -1,5 +1,5 @@
 alter table developers
-add salary double;
+add salary decimal;
 
 SET SQL_SAFE_UPDATES=0;
 UPDATE developers
